dsp: add reciprocal_pll
